biobird_agent/requirements.txt
streamlit>=1.32.0
requests>=2.31.0
python-dotenv>=1.0.0
pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
